+++ R/utils.R
@@ -306,17 +306,26 @@ read_calendars <- function(file) {
     return(rjd3toolkit::.jd2r_calendars(jspec))
 }
 
-#' Write a Calendar file
+#' @title Write a Calendar file
 #'
 #' @description
 #' The calendar file is a xml file like the one JDemetra+ would write when
 #' defining a calendar in the Graphical User Interface.
 #' Calendars can be defined with `rjd3toolkit::national_calendar`
 #'
-#' @param calendars list of calendars or a `JD3_CALENDAR` object
+#' @param calendars named list of calendars or a `JD3_CALENDAR` object
 #' @param file xml format
+#' @param verbose Boolean indicating whether to print additional information.
+#'   Default is `TRUE`.
 #'
 #' @returns \code{NULL} returned invisibly
+#'
+#' @details
+#' If `calendars` is a single calendar (`JD3_CALENDAR` object), it will be
+#' named `cal` by default.
+#' If `calendars` is a list of calendars (`JD3_CALENDAR` object). Then they all
+#' must be named. Else, a error is risen.
+#'
 #' @examplesIf rjd3jars::check_java_version(silent = TRUE)
 #' library("rjd3toolkit")
 #' BE <- national_calendar(list(
@@ -337,8 +346,11 @@ read_calendars <- function(file) {
 #' write_calendars(BE, file = calendar_path)
 #' write_calendars(list(BEL_cal = BE), file = calendar_path)
 #' @export
-write_calendars <- function(calendars, file) {
+write_calendars <- function(calendars, file, verbose = TRUE) {
     if (inherits(calendars, "JD3_CALENDAR")) {
+        if (verbose) {
+            message("The calendar will be renamed `cal`.")
+        }
         calendars <- list(cal = calendars)
     } else if (
         !(is.list(calendars) &&
@@ -360,6 +372,7 @@ write_calendars <- function(calendars, file) {
         jcal,
         as.character(file)
     )
+    return(invisible(NULL))
 }
 
 #' @title Read auxiliary regressors file