+++ NEWS.md
@@ -7,9 +7,15 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+* New `verbose` argument in `write_calendars` and document behavious when the list of calendars is not named [#92](https://github.com/rjdverse/rjd3workspace/issues/92)
+* New message in `write_calendars` when the single calendar is not named (and renamed in `"cal"`) [#101](https://github.com/rjdverse/rjd3workspace/issues/101)
+
 ### Fixed
 
 * Relative paths are accepted for `regarima_read_spec()`, `tramo_read_spec()`, `tramoseats_read_spec()` and `x13_read_spec()` [#91](https://github.com/rjdverse/rjd3workspace/issues/91)
+* `save_workspace` generates a warning when a workspace already exists and `replace = FALSE` (the default)  [#105](https://github.com/rjdverse/rjd3workspace/issues/105)
 
 
 ## [3.8.0] - 2026-07-17
@@ -179,8 +185,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 * New JARS related to version [3.1.0](https://github.com/jdemetra/jdplus-main/releases/tag/v3.1.0)
 
-[Unreleased]: https://github.com/rjdverse/rjd3workspace/compare/3.8.0...HEAD
-[3.8.0]: https://github.com/rjdverse/rjd3workspace/compare/v3.7.1...3.8.0
+[Unreleased]: https://github.com/rjdverse/rjd3workspace/compare/v3.8.0...HEAD
+[3.8.0]: https://github.com/rjdverse/rjd3workspace/compare/v3.7.1...v3.8.0
 [3.7.1]: https://github.com/rjdverse/rjd3workspace/compare/v3.6.0...v3.7.1
 [3.6.0]: https://github.com/rjdverse/rjd3workspace/compare/v3.5.1...v3.6.0
 [3.5.1]: https://github.com/rjdverse/rjd3workspace/compare/v3.5.0...v3.5.1